Move legacy Group back in original folder
