--- packages/flame/src/Group/index.tsx
@@ -1,13 +1,0 @@
-import { InputGroup, SpacedGroup, InputGroupAddon, BeThereOrBeSquare } from './Group';
-import { Group, GroupAddon, GroupAddonProps } from './legacy/LegacyGroup';
-
-export {
-  InputGroup,
-  SpacedGroup,
-  InputGroupAddon,
-  BeThereOrBeSquare,
-  // Here be legacy components...
-  Group,
-  GroupAddon,
-  GroupAddonProps,
-};
