+++ packages/flame/src/Groups/index.tsx
@@ -1,0 +1,3 @@
+import { InputGroup, SpacedGroup, InputGroupAddon, BeThereOrBeSquare } from './Groups';
+
+export { InputGroup, SpacedGroup, InputGroupAddon, BeThereOrBeSquare };
